More sophisticated locking mechanism for shared I2C
Previously, only starting a DMA required a lock of the bus, this is
obviously a bug and causes issues when another task acquires the bus
while a DMA transfer is ongoing.

We have switched to using a type of AtomicDevice which lets us retain
the lock during DMA while allowing interrupts. Many caveats are
introduced.

+ A bunch of other misc fixes that have been waiting to be pushed